requirements.txt
Django~=4.0.4
pillow
requests~=2.27.1
python-dotenv~=0.20.0
requests